move IsEmpty extension from Extensions project to Core project
