+++ src/Core/Extensions/EnumerableExtensions.cs
@@ -12,6 +12,12 @@ namespace AndcultureCode.CSharp.Core.Extensions
         public static bool IsEmpty<T>(this IEnumerable<T> source) => !source.Any();
 
         /// <summary>
+        /// Determines if the source list is empty
+        /// </summary>
+        public static bool IsEmpty<T>(this IEnumerable<T> source, Func<T, bool> predicate) =>
+            !source.Any(predicate);
+
+        /// <summary>
         /// Determines if the source list is null or empty
         /// </summary>
         public static bool IsNullOrEmpty<T>(this IEnumerable<T> source) =>
